// src/models/api-endpoints.ts
export enum VaccineUrl {
  SearchByDist = "https://cdn-api.co-vin.in/api/v2/appointment/sessions/public/calendarByDistrict",
  findByDist = "https://cdn-api.co-vin.in/api/v2/appointment/sessions/public/findByDistrict",
  StateList = "https://cdn-api.co-vin.in/api/v2/admin/location/states",
  SearchBypin = "https://cdn-api.co-vin.in/api/v2/appointment/sessions/public/calendarByPin",
  DistrictList = "https://cdn-api.co-vin.in/api/v2/admin/location/districts",
}